Add Custom Error Matcher for Member Not Found Errors
Summary:
### Summary

We want to explore leveraging Custom Errors more. As part of this we want to add support for matching against our Member_not_found errors.

This diff introduces a custom error matcher for member not found errors in Hack. It adds a new pattern `Member_not_found` to the `patt_error.mli` and `patt_error.ml` files, which allows users to specify a pattern for the class name, member name, and visibility of the member not found error.

The diff also includes a new test runner `test_runner.ml` that verifies the matching of the `Member_not_found` pattern. The test runner checks that the pattern matches correctly for instance methods and properties.

Additionally, the diff updates the `custom_error_eval.ml` file to include helpers for matching the static pattern of the `Member_not_found` pattern.

I also switched test.ml to use OUnit since I couldn't figure out how to have inline tests run with BUCK.

Differential Revision: D81171080

fbshipit-source-id: 76d68f556f60bbe5b8251dfa5e2eeb6e6f6f8138