let integral_hasher be avalanching over all types
Summary:
`integral_hasher` would not be avalanching over 32-bit or smaller types when `size_t` is 64-bit. Change the algorithm so that it is avalanching.

Note that hashers in general cannot be assumed to be avalanching. But there is little harm here. And this may accelerate `F14Table` operations on hashtables with narrow-integer key-types like `int` by allowing `F14Table` to avoid the post-hash mixing step.

Note that this changes the output of integral_hasher over integer types smaller than 64-bit when `size_t` is 64-bit, such as on 64-bit linux platforms.

Reviewed By: Gownta

Differential Revision: D77839012

fbshipit-source-id: deb75ad4e64625dc827e590e2ceb51c09240e04c