Replace usages of mustache last? for unstructured annotations
Summary:
In preparation for moving everything related to unstructured annotations to Whisker, we need to replace usages of Mustache `last?` with `array.enumerate`'s `with_last`.

In the case of Python, the `last?` usage was only for new-line formatting, which is easily replaced without `last?` and makes the output formatting more readable by placing each dict entry on its own line.

Reviewed By: iahs

Differential Revision: D81532786

fbshipit-source-id: 749dac88a615371ed4c2e9068f0917dcc1819a22