IoUringBackend: support multiple provided buf rings
Summary:
Each provided buffer ring only supports 2^15 (32768) entries, which may not be sufficient for an io_uring backend that has tens of thousands of sockets.

Add support for multiple provided buf rings that are handed out to callers round robin. Each provided buffer ring is tightly bound to an AsyncIoUringSocket, until it becomes empty, after which the socket will re-submit a recv.

Reviewed By: yfeldblum

Differential Revision: D81375737

fbshipit-source-id: 1087c02e58f068069c392580f0e71e4491103bd8