Don't lose named arguments when converting nast to tast
Summary:
## What

Fix place where we were losing named arguments when converting the named abstract syntax tree to the typed abstract syntax tree.

## Why

Pre-req to type-checking function calls to functions with named parameters.

## How

Remember names for named arguments.

The solution is just a little bit awkward because the code before was relying on a symmetry between arguments and parameters, which we no longer have.

I actually think the symmetry was more of a coincidence, and it's best to avoid putting redundant data in the AST just to keep the symmetry.

Reviewed By: andrewjkennedy

Differential Revision: D81687433

fbshipit-source-id: 860d030c6494d60ddebd2892f7a6470f19367801