Avoid creating the verifier if we have a default cert verifier.
Summary:
In Fwdproxy, we're creating multiple connections that we're not pooling. We're going to add connection pooling but in the meantime I still think it's worth fixing the bug in this diff. The `HTTPClient` allows you to specify a default verifier but it still creates the verifier first only to throw it away.

This commit changes the verifier creation so that it will not create one if we want to use a default verifier.

Reviewed By: hanidamlaj

Differential Revision: D81638455

fbshipit-source-id: f1ac17fa80632b284e77ed812134926a9d1c414b