Do not pessimize return type inference if we did not intercept
Summary:
Pass a flag to the return type helpers indicating whether we are pre- or
post-intercept. If a function is interceptable, but we are handling a situation
that assumes we are already inside the function, there's no need to pessimize
the return type.

This doesn't do much by itself, but is needed for the next diff that relies on
the type to be at least as good as the specified type hint for builtins.

It will also be useful at some point when we try to inline interceptables in
sandbox mode.

Reviewed By: ricklavoie

Differential Revision: D81721566

fbshipit-source-id: 49ab18d872abe4b1d233704dda79b270e492277d