Update Upper Bound for stack used by FiberManager on TSAN
Summary:
Update the upper bound for stack used by FiberManager on TSAN from 5,000 bytes to 5,632 bytes (5.5 KB). Currently, it's consuming 5,144 bytes.

The upper bound is arbitrary and the stack memory use will ebb and flow as the code changes.

TSAN is documented to increase memory usage by 5x to 10x as it adds additional function calls on data accesses & function entry/exit[0][1][2].

Also, add back the seperate limit for when TSAN is not used (TSAN was a sepearate block before D24179337).

0: https://clang.llvm.org/docs/ThreadSanitizer.html
1: https://llvm.org/devmtg/2012-11/Serebryany_TSan-MSan.pdf
2: https://www.youtube.com/watch?v=HDgttiIvMxA

Reviewed By: yfeldblum

Differential Revision: D81701155

fbshipit-source-id: 3cdd20dc18496cdac1c5a6e55e3187542c52a02f