Add a Benchmark to Measure Context Pool Performance
Summary:
Per Orvid's comment on D64837870, it would be interesting to move to
dynamically sizing the core local context pool cache. I am interested in
exploring what that sizing policy should be. In order to do so, I want to be
able to measure the performance of the pools under different scenarios.

In the past I've done so by measuring overall compression performance of
small inputs but that's a pretty crude method. This is more direct and more
sensitive. Although of course it is also more artificial.

Reviewed By: yfeldblum

Differential Revision: D64973230

fbshipit-source-id: b962eb9ce65dd3eb4e77b19b69d003bf31fca1d1