Log when the unit of a softly deployed file is merged
Summary:
A slightly different take on {D65793639}.

If a file is softly-included in a build, then we're attempting to remove it from the deployment and there will be Hack errors on static references to its declarations. However, we also want to catch any dynamic loads of the unit (similar to __Deprecated) and then hard exclude when we're free of logs. Soft inclusion is already tracked on unit-emitters, but the logic was the old modules-based package v1 logic. Update it, computing when finalizing unit emitters, and log when merging units.

Reviewed By: sinancepel

Differential Revision: D81289573

fbshipit-source-id: 613f1a815f8702ecd168176d638134595f68b707