Explicitly declare/define the specializations of `patchByKeyImpl`.
Summary:
This addresses a known issue with modules where using a template that is undefined at the use-site can break.
Specifically, given something like:
```cpp
// a.h
template <typename T> void f();

void g() { f<int>(); }
```
```cpp
// a.cpp
#include "a.h"

template <typename T> void f() {}
```
While this pattern is certainly unorthodox, compiling `a.cpp` was not a problem before.
Since the point of instantiation of function templates is at the end of the translation unit,
`f<int>` was instantiated at the end of `a.cpp`.

This breaks with modules however, since `a.h` is considered the translation unit, `f<int>` is instantiated at the end of `a.h`, without the definition being visible.

This diff fixes this problem by providing explicitly declarations (in the header) and definitions (in the source).

Reviewed By: vitaut

Differential Revision: D81730918

fbshipit-source-id: 683228bc20f1fbb03074ee177ad1c624f2321118