Remove invalid assert about existence of type of pointee
Summary:
As demonstrated by the unit test (-r --jit-serialize 2), we could profile that
we likely have a struct dict containing an object, but actually have a static
dict (which can't contain an object) during inlining. We still emit a guard
based on the profiling data (because that logic is not smart enough to figure
out it is an impossible situation) and then later realize that the result of
LdMem of StructDictElemAddr must be TBottom.

Even if it is possible to teach JIT that struct dicts with TCountables must be
TCountable themselves, there is nothing that generally guarantees that the
logic inside typeOfPointeeFromDefs() won't infer more info and produce a
TBottom.

So let's remove the assert.

Reviewed By: ricklavoie

Differential Revision: D81979015

fbshipit-source-id: f7f3ed2f774ba07b04ec64bf56bea5c25bd5d84a