Add callback setters on the ConcurrencyControllerInterface for expire/execute notifications
Summary:
One of the requirements of mysql with DLS is to get notifications when requests expire/execute so we can generate proper statistics in information_schema tables.

The two existing controllers currently support that (for ODS logging purposes), but not the custom mysql controller that we use to integrate with the CPU schedule.

To enable notifications, I moved the two callback setters to the `ConcurrencyControllerInterface` interface so that in a later diff I can hook up the mysql callbacks.

Reviewed By: tlj77

Differential Revision: D81931440

fbshipit-source-id: 63075a16a2c4a49acec22a01f0d960c863954034