Back out "Ensure Request Context is saved/restored with BlockingWaitExectuor"
Summary:
S557355 Crashes in folly::RequestContext::setContext during shutdown across binaries SEV3

Increase in crashes seems to align with when this diff was committed.

As this change was not required at the time, and only fixed a potential issue, it should be safe to backout.

We aren't sure if this will fix the issue, but since this is a relatively safe diff to backout, we can remove it and see if the problem is resolved.

Reviewed By: ispeters

Differential Revision: D81956075

fbshipit-source-id: 7718f87a1eadb12f06224568d611328732b305ec