Do not include unnecessary headers in patch
Summary:
Today if we have

```
typedef foo.Foo Foo
```

We will import `foo.thrift` in python patch generated code, thus we will need to generate patch for `foo.thrift`.

This is wrong since we can only patch struct/union, we can't patch typedef by itself.

This is also inconsistent with BuildGraph API.

This diff changed so that we only check structured for files that need to be imported.

Reviewed By: alexq-fb

Differential Revision: D81982761

fbshipit-source-id: 70a56f3d1930e11e8427bd4c56902e081b64d9b2