add jk to disable include_path strict parsing
Summary:
The flag `~strict` of `PackageConfig.load_and_parse` controls whether hh should fatal at startup whenever PACKAGES.toml defines an `include_path` that does not exists on the filesystem.  The strict flag was hardcoded to `true` for hh (it is set to `false` for hh_single_type_check as multifile unit tests trigger false positives otherwise).

This diffs adds a local config to override the `~strict` flag in hh.

Reviewed By: DavidSnider

Differential Revision: D82015759

fbshipit-source-id: c03dc3c25c821ca66d083439485040b90f866928