SmallSerialExecutor -> SerialExecutor
Summary:
`UnboundedQueue` is not a good fit for `SerialExecutor`, due to its large memory usage, while `SerialExecutor` is supposed to be lightweight so that it can be used at request or even finer granularity.

`SmallSerialExecutor` has been adopted for several use cases and it has shown comparable performance while being much more memory efficient. This gives enough evidence that we can retire the `UnboundedQueue` version and make `SmallSerialExecutor` the default.

Reviewed By: Gownta

Differential Revision: D81315614

fbshipit-source-id: 73e5e013539b7551f029579362b351608f2dfcde