Add Whisker parsing validation test for all compiler templates
Summary:
Add a parameterized test to check that all Whisker templates bundled into the Thrift compiler are valid.

This is motivated by D81830547, which allowed a template with `{{#each expr}}...{{/each expr}}` to land. While `{{#if expr}}` *requires* the condition expression to be present in the closing tag if the closing tag is on a different line, it is *always invalid* to close `{{#each expr}}` with `{{/each expr}}` - it must always be closed with `{{/each}}`.

This test should catch this entire class of errors (invalid Whisker template) in the dev inner loop, even if a particular template is not evaluated by our other test cases (e.g. fixtures don't trigger conditions guarding a macro inclusion).

Reviewed By: praihan

Differential Revision: D81985935

fbshipit-source-id: 4327408b280717ccc347692cb16fb844e8105c77