Remove an #include that is not used.
Summary:
VS Code reports this is not used, and it doesn't appear to be. Consumers should presumably be using this explicitly, as using C-style includes has gone out of fashion.

Note `#include <new>` is also "unused" and could also be removed in a later change.

(I am new to Meta and am doing this as an experiment.)

Reviewed By: Gownta

Differential Revision: D81962310

fbshipit-source-id: a5e119a90da71cff83b5039dec79b5df2d262365