add tests to show behavior of __Const property object getting mutated
Summary:
What does it mean for a __Const property to be of type `class A`? PHP
readonly properties will allow for interior mutation [1] and so does Hack. For
now, I'll add a test that proves this. Should we decide that const properties
mean properties are frozen, I'd like to have this catch those changes.

[1] https://www.php.net/manual/en/language.oop5.properties.php#language.oop5.properties.readonly-properties

Reviewed By: dlreeves

Differential Revision: D82047609

fbshipit-source-id: 2613263996c1e3d9188c5b868c80b9b9e0281dc4