Add rust_fields_reversed function to Thrift Rust generator
Summary:
This adds a new mustache template method "struct:fields_reversed" that returns struct fields in reverse order. This is useful for generating Rust code that needs to process struct fields in reverse order. Which is a key part of the "fast path" implementation used in the next diff in this stack.

The implementation follows the same pattern as the existing rust_fields_by_name method:
- Copy the struct fields using `struct_->fields().copy()`
- Reverse the order using `std::reverse(fields.begin(), fields.end())`
- Return them via `make_mstch_fields(fields)`

Usage in mustache templates:
```
{{#struct:fields_reversed}}
  // Process fields in reverse order
{{/struct:fields_reversed}}
```
 ---
> Generated by [Confucius Code Assist (CCA)](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=2fa66994-5d28-11f0-9d3e-69474d1a5a72&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=2fa66994-5d28-11f0-9d3e-69474d1a5a72&tab=Trace)

Reviewed By: dtolnay

Differential Revision: D78057089

fbshipit-source-id: 32f9b2abebb2d3b9509e4e2175a20b19b949f42f