Do not claim that calls or reentrance can modify arrays and closures
Summary:
It is impossible for a callee (or VM reentry) to modify an array that is being
tracked across the call. Or a closure, as they are const.

If an array is tracked across a call, the caller must hold a reference to it.
Otherwise, if the callee "modifies" it, if the refcount was 1, it could trigger
a reallocation, making the caller's SSATmp invalid, which would crash HHVM.

This is true even if the reference to the array is stored somewhere else (e.g.
a property). If the same location (e.g. a property) is being used before and
after a call, it needs to be reloaded, resulting in a new SSATmp. Array element
"modifications" are irrelevant, what has changed is the property (even if the
modification happened in-place and the array pointer remains the same).

This diff removes AElemAny and AClosureArgAny from all store memory effects
that relate to execution of random Hack code (but keeps it for array and
closure specific instructions).

It also preserves information about array element locations in load-elim across
function calls.

Reviewed By: paulbiss

Differential Revision: D69281340

fbshipit-source-id: c355cc3763d487810aebda31dfa2eac92f9557fc