extend ServerNotifier telemetry
Summary:
This diff extends the telemetry we collect for calls into the file watcher, for both Eden and Watchman. Currently, all collection of telemetry for this happens in `ServerMain.query_notifier`, which is the main point where the server calls into `ServerNotifier`.

This diff adds some telemetry collection to `ServerNotifier` itself, measuring the duration of just the calls to `Watchman.get_changes*` and `Edenfs_watcher.get_changes*`. The idea is to get more accurate results on the duration of the request to the file watchers themselves, without including any of the other things that `ServerNotifier` may do. For example, `ServerNotifier.get_changes_{sync, async}` calls `ServerRevisionTracker.files_changed`, which can cause us to run `hg`.

The Eden team is interested in this data, to compare the performance of the file watchers directly.

To implement this, I've created `Telemetry.add_duration` as a counterpart to `Telemetry.duration` which accumulates existing durations with the same key.

Reviewed By: madgen

Differential Revision: D82398694

fbshipit-source-id: 70d566ac67b68400a7bc220e31a74cc3ceeec5fc