fix BenchmarkingStateTest.PerfIntegration flakiness
Summary:
The test seems like it should be deterministic but is observed to be flaky.

```
[ RUN      ] BenchmarkingStateTest.PerfIntegration
fbcode/folly/test/BenchmarkTest.cpp:275: Failure
Value of: perfOuptut
Expected: has substring "Performance counter stats for process id"
  Actual: ""

[  FAILED  ] BenchmarkingStateTest.PerfIntegration (51 ms)
```

Hypotheses: The `perf` subprocess is sent `SIGINT` before it gets a chance to attach to the parent (this), which means it does not ever emit output, leading to flakiness.

Solution: Wait for the `perf` subprocess to attach to the parent. Use `perf_event` file descriptors as signal.

Reviewed By: DenisYaroshevskiy

Differential Revision: D82406891

fbshipit-source-id: 372b93eab4d1c31bd24485b5ef113cdc0cfd362d