Improve behavior of RequestContextSaverScopeGuard
Summary:
The main use case for `RequestContextSaverScopeGuard` to guard a loop is supposed to mimic using `RequestContextScopeGuard` in the body of the loop, but optimizing for the case where consecutive iterations have the same context (we avoid a unset/set pair). However, there is a noticeable difference: when the context is not the same, the old context may be destroyed under the new context, instead of the parent context (which in most executors is usually null).

This is not a major issue, but it can cause unwanted propagation of the request context if some `RequestData` destructor schedules work into an executor.

To fix this, pull the "same context" check inside a guard method, and do the unset/set explicitly otherwise. This also has the benefit of not encouraging people to use `RequestContext::setContext()` directly, which we might be able to make private eventually.

Also add `restoreContext()`, which is used in the next diff.

Reviewed By: yfeldblum

Differential Revision: D82399528

fbshipit-source-id: 2f5ac7586b406855467e4aeb6542b098de4c77f7