Raise notice when non-async methods override async methods
Summary: The last attempt to make this a hard check (D76155725) failed because of violations in the FBMock framework. In the current FBMock implementation, an async method is mocked by overriding it with a non-async method that returns an awaitable. This leads to a violation of the check introduced in this diff. Log violations and fix them before hard enabling this check in the runtime as there may be other such violations in WWW.

Reviewed By: jano

Differential Revision: D76982363

fbshipit-source-id: b40d21227cb02373f906080b6702cdb83026c287