Fix always is/as false positives
Summary:
In order to do subtype redundancy queries, we were replacing wild cards in type tests with fresh type variables. The produced type was also used for the disjointness check that is adjacent to the redundancy check. This was leading to false positives, e.g.,
```
function genBKRootActionWithActionType(dict<string, mixed> $d): void {
  idx($d, 'data') as dict<_, _>;
}
```

In this diff, we change we move the mutation to the type around so that it doesn't affect the disjointness check and hence makes the false positives go away.

While making the changes, I noticed that there was also another source of false positives for redundancy checks in the case of constraints. When the type under test is not solved, the additional `sub_type` constraint was causing it to be solved to `nothing` hence leading to false redundant assertion warnings. We now force solve the type prior to adding the `sub_type` constraint to eliminate those as well.

Reviewed By: geralt-encore

Differential Revision: D82051172

fbshipit-source-id: 18c31114c263a0155af9e9611dd3872b71e2a8e7