set file_watcher in hh_server_events
Summary:
We currently have an integer column `use_watchman` in `hh_server_events`, which tells us if `hh_server` is using Watchman.

This diff generalizes that, by introducing a new column `file_watcher`, with string values "dfind", "watchman", or "edenfs". For backwards compatibility, I'm still populating the `use_watchman`column as before. Longer term, we could make that a derived column.

Having this new column is useful due to the following:
The information in the `rollout_flags` column can only tell us whether `Edenfs_watcher` *should* be enabled, but it doesn't tell us whether it actually *is* enabled: If `Edenfs_watcher` fails to initialize (for example, because we are not on an Eden-backed www repo), it falls back to Watchman instead. I do log whenever that happens as a `EDENFS_WATCHER_FALLBACK`event, but it seems sensible that each row should tell us more easily what file watcher is used.

Reviewed By: madgen

Differential Revision: D82302010

fbshipit-source-id: ba78162364f67b05e3e2cf33bc2f707061132352