Migrate template usages of interaction cpp_name to interaction name
Summary:
In the before state, `service:cpp_name` has a specialized implementation where if the `service` object is an interaction, `cpp_name` is just the normal name.

This change moves that logic into the templates - using `interaction:name` for interactions and `service:cpp_name` for services. Once we're confident, we can remove the specialized implementation of `service:cpp_name`.

Reviewed By: iahs

Differential Revision: D82654551

fbshipit-source-id: e130f57b134de65c2fda9f2ba6fbf91e7fe7b076