make hh --find-my-refs walk call-graph up to max distance
Summary:
This diff allows `hh --find-my-tests` to walk reference edges backwards (i.e., from referent to referee) up to a maximum distance in order to reach test files from the set of initial symbols.

The implementation is still limited to methods, but most pieces are in places to extend it to other kinds of definitions.

A downside of the current implementation is documented on `enclosing_def_of_symbol_occurrence`: It relies on `FindRefsService` (which in turn uses `IdentifySymbolService`), which means that we only the source positions of any referees to the current method, but no information about the enclosing definition (which we need to process next). This means that we currently traverse the NAST of the file in question to find the enclosing method ourselves.

Reviewed By: madgen

Differential Revision: D81884479

fbshipit-source-id: dc4530ed5f15aa8aacfac8a7e2d82cc812319b7b