hh --find-my-tests
Summary:
This diff adds a new subcommand to `hh`'s check mode: `--find-my-tests`

The motivation is to replace the part of `FindMyTests.php` that actually searches for tests. Moving the test selection into `hh` (rather than needing to call `hh --find-refs`, and then process the result), means that we are closer to the source of truth.

In its current form, the implementation is mostly just using the logic of `--find-refs`, but
- only works on methods
- but accepts more than one symbol as input
- only returns the paths of the files that have references, without the name of the actually referenced symbol, or further position information
- only returns files that are tests (determined by whether they appear under a `__tests__` folder)

Note that this means that this command implements none of `FindMyTests`s processing of diff hunks.

Implementation note: A lot of the new code in `serverFindMyTests.ml` is taken from `serverFindRefs.ml`. I didn't want to just expose those parts in `serverFindRefs.mli`, I expect that code to change soon. `ServerFindMyTests` does rely on `FindMyRefsService` (different module!) public API, which seems fine.

Reviewed By: madgen

Differential Revision: D81882314

fbshipit-source-id: 6623a2ab9bce243471f06d0e883949293cd7ac72