Simplify to a PyCapsule we can easily and cheaply use
Summary:
Moved to storing the request context directly in a PyCapsule in the ContextVar is makes it easier to use when integrating with other c-api code that might need to access it.

Fixed bug where we were not properly clearing the RequestContext when the PyContext didn't have one set, otherwise RequestContext could leak into other context that didn't/shouldn't have a set RequestContext.

Removed some interfaces that were not really needed at runtime/userland and could be dangerous.

Reviewed By: itamaro

Differential Revision: D82676452

fbshipit-source-id: fd711a47284d54bf1a1fd6ce52133995c9e7245f