Introduce folly::irange
Summary:
A type-safe, const-safe for-loop.

This has nicer ergonomics than `std::views::iota` in two ways. First, there is
a one-argument version `irange(N)` that loops through the range `[0, N)`
whereas `iota(N)` will loop through the range `[N, max/infinity/???)`. Second,
if the data types of two input the arguments mismatch the first argument will
be cast to the type of the second argument with a sanity check (which could be
strengthened).

swolchok and I introduced this to PyTorch and ExecuTorch a couple of years ago
and it is now the primary and recommended way of writing host code for-loops in
those projects. It prevents issues such as
```
for(int i=0; i < vec.size();i++)
```
where large datasets can result in indexing issues due to range limitations on
`i`.

Reviewed By: yfeldblum

Differential Revision: D82760961

fbshipit-source-id: f675868d51b9093007d03b859a4255a50ac08b26