Include IC with Xenon results
Summary:
MCP has been moving towards using the IC to do cost attribution. This allows propagating the "owner" of the compute to be propagated across services.

We currently only use set_frame_metadata for Xenon, even though the IC propagate works much better and is already set.

We're hoping that exposing the IC to xenon will fix a few cases of blame:
- Unawaited-awaitables
- Some batching cases which specifically use `embed_implicit_context_state_in_async_closure`
- PSP stitching sometimes fails, even for Xenon, but PSP IC propagate never "fails".

The feature works by serializing the IC state and embedding it in the xenon results similar to other results. This is somewhat non-ideal because it exposes the internal IC state without using the wrapper classes, but I figure since it's for a profiler, this is fine.

We considered an alt approach where we'd just allow a callback to be registered to Xenon (similar to fb_setprofile), but this would forever make Xenon re-entrant.

Reviewed By: jano

Differential Revision: D82354347

fbshipit-source-id: 567d438e8dbe567885fc22c6d0034ebc7e239052