Make ApplicationExceptionIf extremely unique
Summary:
Assertions are very brittle at the moment.

For example - `ProtocolException` asserts to `TransportException` - that's a disaster waiting to happen:
https://www.internalfb.com/code/fbsource/[7b7fd7ac7d71]/fbcode/thrift/lib/go/thrift/types/protocol_exception.go?lines=24-27
https://www.internalfb.com/code/fbsource/[7b7fd7ac7d71]/fbcode/thrift/lib/go/thrift/types/transport_exception.go?lines=29-33

I need to avoid the same destiny to `ApplicationExceptionIf` - by carefully eradicating it downstream.
To avoid breakages during the eradication process - I need to make the interface "extremely-unique" to avoid false-positive assertions.

Reviewed By: mahmednabil109

Differential Revision: D81892797

fbshipit-source-id: 5b89d14d3e484bde0de1e26b291fa932cbf1d290