Do not change dynamic parameter type with type of default expression
Summary:
Currently, to check a function such as `function foo(t $x = e)` we infer a type for `e`, and check that the type of `e` is a subtype of `~t`. If `t` is not enforceable, the type assigned to `$x` is the union of `t` and the type of `e`, as though there was a conditional `if (argument_missing) { $x = e; }`. But if `t` is enforceable, then we not that the declared type is preserved.

Unfortunately, if `t` is `dynamic`, and `e` has type `~int`, say, we end up checking the body assuming that `$x:~int`, which leads to downstream errors if `$x` is passed to something expecting `dynamic` or another like-type. So we should make a special case for `dynamic`.

Reviewed By: vassilmladenov

Differential Revision: D82964709

fbshipit-source-id: 1d1e0b0ad55f0680f72aee73edd342db73269d85