Fix OPAM build for Arm
Fetch the arm64 version of opam for installing deps,
and disable the `ocaml-option-fp` flag which is not supported on aarch64.

The pinned versions of `dune` and `landmarks-ppx` still need updating
to work on aarch64. This will be done as a followup since it seems to be
causing issues in internal CI at the moment.