Eliminate promise-punning UB from getErrorHandle
Summary:
As discussed in the preceding D82907354, it is quite hard to eliminate promise-punning UB from `TaskWrapper::get_return_object`, so we just leave it be with the safeguard of `is_promise_type_punning_safe`.

However, there's no inherent reason for `getErrorHandle` to interact with `coroutine_handle<Promise>`. It actually only either needs the `Promise&` (which may be stored as a member of a wrapper promise without issue), or the type-erased `coroutine_handle<>`. Furthermore, the latter is only needed on the "no-op" path, i.e. the per-coro implementation does not need to see a handle at all.

This diff renames `getErrorHandle` implementations to `getErrorHandleImpl` for clarity, and switches them all to the new, UB-free design.

Reviewed By: ispeters

Differential Revision: D82926069

fbshipit-source-id: 3230102ffa12ae466ca822ce6a058092c720ac5d