Refactor helper methods on ServerGeneratorStreamBridge
Summary:
Make helper methods in `ServerGeneratorStreamBridge` static to indicate they do not depend on the bridge's state.
This diff also puts ContextStack pointer back to the `fromAsyncGeneratorImpl()` method, as that is necessary for the future refactoring where we want extract static methods (`fromAsyncGenerator()` and `fromAsyncGeneratorImpl()` completely out of the `ServerGeneratorStreamBridge`).

Reviewed By: praihan

Differential Revision: D82694990

fbshipit-source-id: 891acfc04fb87bb7f4579d3ddf367bd90f07acda