completions for named params
Summary:
Make named parameters work better with completions.

## Note testing for fake arrows

> The term "fake arrows" refers to how we, for example, offer `C\count($x)` as a completion for  `$v->count` when `$v` is a collection.

I implemented support for completions for named parameters for fake arrows just in case somebody adds something to `C\` or `Str\` that has named parameters.

In order to test this, I had to add an imaginary standard library function that won't show up in prod (because not in the naming table). I don't have a strong opinion about whether or not to leave the fake function in.

Reviewed By: geralt-encore

Differential Revision: D82231990

fbshipit-source-id: 0ba80b5cab8630c3b3e9820b1f755053cac4f88b