show bug in decl_fun_utils
Summary:
## What

Show minor bug in the existing check for forbidding required params after optional params.

The function is documented as producing an error on the *first* non-optional parameter after an optional parameter:

https://www.internalfb.com/code/fbsource/[d890a168cb3f17282e4003eb8b3a3b97f376b439]/fbcode/hphp/hack/src/decl/decl_fun_utils.ml?lines=117-120

However, it actually does something weirder, since the check resets state after each optional parameter, erroring on *both* `$x2` and `$x4` in the following case, but not on `$x5`:

`function f(int $x1 = 1, int $x2, int $x3 = 3, int $x4, int $x5): void {}`

## Why

We change behavior and refactor later in the stack, in order to facilitate easy-to-maintain and accurate checking in the presence of named parameters.

Differential Revision: D83053409

fbshipit-source-id: 08faafcb150ab290ec9c500245928b439888e18c