In ROAR, try ROAR symbolizer first
Summary:
For a given JIT address, trying to symbolize it with
file->getDefinitionByAddress is a waste of cycles. For a given no-JIT
address, trying to symbolize it with ROAR is also a waste of
cycles. It all comes down to how fast each technique can discard a
sample and what is the ratio of JIT over no-JIT symbols. For AdFinder
workloads, which use the symbolizer a lot, most of the addresses are
JIT addresses, and we are wasting a considerable amount of cycles on
calling file->getDefinitionByAddress first only to discover later that
this is a JIT symbol.

Switch to call ROAR first, which should discard no-JIT address as soon
as a series of comparisons against arena boundaries determine that
this address is not in any part of the JIT. This will replace the
current approach of invoking an iterator on ELF sections for each JIT
address.

Reviewed By: jpporto, NTillmann

Differential Revision: D83170135

fbshipit-source-id: 1bb93f188f5e6dbe162758787d142429dfb29118