document divergence between typechecker and runtime for Const prop
Summary:
`__Const` properties should be semantically equivalent to `readonly`
props in PHP or C#. C# declares the following:

> A readonly field can't be assigned after the constructor exits. This rule has different implications for value types and reference types:
>  Because value types directly contain their data, a field that is a readonly value type is immutable.
>  Because reference types contain a reference to their data, a field that is a readonly reference type must always refer to the same object. That object might not be immutable. The readonly modifier prevents replacing the field value with a different instance of the reference type. However, the modifier doesn't prevent the instance data of the field from being modified through the read-only field.

https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/readonly

The runtime currently enforces that const props cannot be mutated, and
respects these two rules. Observe in the HHVM tests that a value type property
cannot be mutated even by a method call.

The typechecker does not respect this for the postfix integer increment call,
either via a method call or a direct call.

For now, this tests just documents the fact that integer increment does not
trigger the same codepath as assignment to a field. I'll add a followup that
changes the typechecker to align behavior with the runtime.

Reviewed By: mdko

Differential Revision: D83152233

fbshipit-source-id: 410d2c3f1295f18afffc72dd87a858b21b696197