Fix GADT-style inference for refining scrutinee with with refinements
Summary:
When we do `if ($x is Foo)` where `$x: Bar<T>`, we have the opportunity to learn something about `T` if `Foo` is a descendant of `Bar`.

Assuming that `Foo` can only inherit from `Bar` once (true barring an unsafe attribute to unbans this), we learn something about `T` by adding the constraints inferred from the proposition `Foo <: Bar<T>`.

However, we just use the type of `$x` as the RHS of the subtype proposition, but if the type of `$x` is `Bar` but with with-refinements, then the proposition ends up just being false if `Foo` doesn't match the with-refinements.

Just removing the with-refinements from the super type is sound but not complete since we may fail to learn something about generics use in the with-refinements. (See "gadt_refinement_refines_generic_in_with_refinement.php")

What this diff does is it copies any with-refinements to the sub class type when producing the subtype relationship. This is valid because, in the case that the predicate passes, these with-refinements must be valid for the sub class type since we only produce the relation ship the super class is an ancestor of the sub class. When doing the copying, we have to convert any exact with-refinements to the equivalent loose with-refinements so as not to shadow any existing information about the type constant on the sub class.

> note: I realized when doing this that we only create an uninstantiated subtype prop for this specific case so I've specialized the code around this and removed some superfluous abstraction.

 ---

Note that this only updates the new refinement path. The old path is also susceptible to this too (https://www.internalfb.com/code/fbsource/[069e164a9c0f]/fbcode/hphp/hack/src/typing/typing.ml?lines=2347) but seems more cumbersome to fix.

Reviewed By: dlreeves

Differential Revision: D81258922

fbshipit-source-id: e3e5754d38a360de0bf81523783891bbc050c2f7