Make typing/dune more like typing/BUCK
Summary:
Conform our dune library dependency structure for typing libraries to be closer to that of our buck library structure.

This makes future changes easier, including changes I hope to do later in the stack.

Treating BUCK files as the source of truth for desired library structure, because that's the main thing we use and the way we build when we ship to users. We avoid unintended user-facing effects by not changing BUCK files.

Differential Revision: D83057997

fbshipit-source-id: 21dbb50d35cc0571bfcddd1d6a718f68ad96d64f