Migrate typeinfo template to operate with mstch or Whisker type context
Summary:
Make the `typeinfo` template for the Python generator work with either mstch_type or a Whisker type object as its context.

The template primarily renders with Whisker type as its context. If the context is a mstch object, the template recurses itself using the Whisker `self` property as the context for the recursion.

This enables incremental migration of calling templates from mstch to Whisker, without requiring all templates to be updated to wrap the macro with a `{{#type:self}}` context that would later need to be removed.

Reviewed By: praihan

Differential Revision: D83302449

fbshipit-source-id: 8e3e1f85b55ab547106c1b803359a9f9676f2006