Improve BlockingQueue::size() implementations
Summary:
Getting the queue size can be inefficient, especially for `MPMCQueue`, and priority queues need to add up across priorities.

We can instead rely on the invariant that the number of posts in the semaphore is equal to the number of entries in the queue. All `BlockingQueue` implementations do this because it's needed for dynamic thread pools, and actually the bounded implementations are slightly incorrect and being fixed in this diff.

Also remove the method `sizeGuess()`, which is not part of the `BlockingQueue` interface and it's unused (since `size()` does the same thing). This is probably some leftover that was copypasted into other implementations.

Reviewed By: yfeldblum

Differential Revision: D83350138

fbshipit-source-id: b84c9532e617172cb032a299af5d4759e069af08