do not log `fromIncludeDir` files to scuba package drift
Summary:
Do not log to scuba the "Packagev2 WWW Build Drift" table files included in the repo with a `fromIncludeDir` reason, as these do not belong to an excluded dir and are irrelevant for symbolrefs/packages drift.

Also, following the PackageV1 removal, delete the `V2` suffix from the Package reasons.

Reviewed By: mdko, DavidSnider

Differential Revision: D83342938

fbshipit-source-id: ef942440a5b82d8674d5d13702b8608f0c9c7981