invoke ::handleRequest on ingressErrors
Summary:
## motivation:

Upon ingress errors (think 408 ingress timeout, 400 bad request), we want to allow the application/handler to be able to send custom responses (e.g. revproxy sending custom html/css back to user for DOM based rendering)

## design:

One method to achieve this feature is having a virtual function in the handler specific to ingress errors (for example, this is how legacy proxygen/lib does it – [codeptr](https://fburl.com/code/glsucowv))

Although the above is possible, this solution goes thru the existing `::handleRequest` interface and gives the user a chance to return an http response with the code we expect (e.g. 408, 400). However, its important to note the application/handler **must** send the expected status code in its response, otherwise the session will generate a default response with said status code on behalf of the app/handler. This is debatable, but ultimately boils down to the core principle of never trusting application handlers "to do the right thing". For example, sending a 200 on an ingress parsing error makes no sense, should be disallowed, and will result in a lack of downstream understanding/debuggability.

## changes:

Adds a `uint16_t errorStatusCode_` member to HTTPStreamState representing the expected status code for the application/handler's response. This is set on ingress timeout errors (to 408 via `HTTPCoroSession::sourceComplete`) and ingress parsing errors (to 400 via `HTTPCoroSession::onError`). Once the HTTPHeaderEvent is yielded from the handler's source, we compare against the `errorStatusCode_` – if equivalent we let the source thru, otherwise we abandon reading from the source and override with a session generated source via `::getErrorResponse`

Differential Revision: D77162016

fbshipit-source-id: 2fd0eee56c16e6758491766906019311dac03524