Improve error messages for constraint-based-typing
Summary:
Improve some error messages where the call to subtyping uses the whole call expression as the primary position. However, the error might have a more specific location, like the method name (if it isn't there), or a particular argument (if that's where the error arose). So inspect the goo that is Typing_error.t to detect these cases and update the position. This required adding a With_pos constructor to Typing_error.t, in analogy to With_claim.

There are lots of errors that aren't argument/parameter mismatches: the method doesn't exist, a where constraint was violated, etc. None of these are Invalid arguments, but come from other underlying primary errors, so we remove the top-level of the Error.t in these cases.

Reviewed By: enetsee

Differential Revision: D82646247

fbshipit-source-id: a902f044cf2ef1176c6d5b61bb058ec549d34813