Improve intersection simplification by adding supportdyn from tyvar upper bounds
Summary: Subtyping simplification is incomplete for intersections on the left of a constraint e.g. `#0 & t & u <: CanTraverse(#1)`, because we turn this into a disjunction. In particular, we sometimes "lose" the fact that a type variable is bounded by a type that supports dynamic, and end up with errors such as "dynamic is not a subtype of mixed". (Here, because we elide `supportdyn` when printing, we can't see from the error that the `mixed` is not in fact `supportdyn<mixed>`.) In general, it's sound to propagate upper bounds from type variables onto the other conjuncts in the intersection, because `#0 <: t /\ (#0 & u) <: v` iff `#0 <: t /\ (#0 & u & t) <: v`. We could do this for any bounds on a type variable, but I'm not confident about the performance implications. So for now, we apply some simple syntactic tests on the type variable bounds to check for "supports dynamic", then add `supportdyn` to the other conjuncts in the`simplify_intersection_l` helper. The two new tests demonstrate instances where this removes errors.

Reviewed By: vassilmladenov

Differential Revision: D83279327

fbshipit-source-id: f0e1a29a2b94bb7134db65ca5c208c8c747dadd0