Enable custom folly::not_null failure handling
Summary:
This extends the definition of `folly::not_null` to support alternate behavior when throwing exceptions on null assignment/construction or in debug mode when the not null invariant condition is violated and the application should be terminated. This is leveraged in other changes to define an alternate in a codebase that has historically favored use of `gsl::not_null` that treats null assignments/constructions also as logical errors violating the invariant condition and triggers an app crash. This is viewed as desirable as it makes it easier to detect and correct such logical errors as they will trigger crashes with the call stack and context reflecting the point of the violation, rather than throwing an exception which either the program will attempt to handle or may trigger a failure elsewhere after the original context has been lost. For argument validation at API entry points, explicit checks are preferred as opposed to using `not_null`. This opens up use of a `not_null` wrapper with smart pointers (which `gsl::not_null` doesn't really support, particularly due to move semantic support) but with more strict null handling behavior that can also include things like logging of failures.

Feature code changes:
* Add an additional template argument to `not_null_base` and `not_null` that defines the type containing the null handler behavior functions. This approach should mostly eliminate any binary size growth relative to the original implementation if only one handler specialization is used and doesn't have any virtual function dispatch overhead.
* Add default null handler implementation that retains original behavior.
* Add explicit default `not_null` destructor to silence linter warning.
* Add some more explanatory comments both for existing behavior and new behavior details in regards to handling of null handler arguments.
* Remove unused predeclaration in fbcode/payment/basedb/Connection.h that caused build errors due to missing new template argument.

Test code changes:
* Updated existing unit tests to be type parameterized on the null handler so that they can be specialized to run on both the default null handler and a test alternate handler.
* Add some type aliases to fixtures and test instances to cut down on some duplicative template boilerplate.
* Add many new tests that validate conversion between `not_null` specializations that use different null handlers.

Reviewed By: billorr

Differential Revision: D83199668

fbshipit-source-id: 3b22d5a9fedcf273c33758247a474391dfed6e36