correct function names: s/variadic/splat
Summary:
## What

Rename internal functions to use "splat" terminology instead of "variadic"

No change in user facing behavior.

## Why

iuc, "variadic" is a declaration-side concept and "splat" refers to something that can occur in an argument list. So there's no such things as a "variadic expression."

## References

I couldn't find any mention of `...` in hacklang.org.
PHP docs refer to it as "the `...` operator ... also known as the splat operator in other languages ..."

https://www.php.net/manual/en/migration56.new-features.php. Hack docs clearly refer to `...` in declarations as "variadic": https://docs.hhvm.com/hack/functions/introduction#variadic-functions.

Differential Revision: D82307547

fbshipit-source-id: 7e63c61ea3d043c91981e5f8897a983cd5b01bd4