provide a get_PyContext that can be used with current context or passed in context
Summary:
I saw in the thrift-python for its ContextVar that it sets as part of RPC's it makes the context var only accessible from cython by storing it in a `cdef object` which keeps it out of the module dict and no longer accessible from python code directly.   Not since we only want to maintain that PyCapsules are stored in the ContextVar we need to lock down the var.

This creates a get_PyContext which can be used to operate on the context var without the ability to mutate it in "undesireable" ways.

Reviewed By: yfeldblum

Differential Revision: D83683929

fbshipit-source-id: 4b765940be31c7f65523166e598e986a6a12cb39