distance 0 selection
Summary:
This diff adds special handling for the case where a symbol passed to `hh --find-my-tests` is inside a file that contains test classes. In this case, we immediately select it (with distance 0).

Technical details:
Given that `hh --find-my-tests` is just doing a BFS traversal of the call graph at the moment, one may expect that we would handle this case (i.e, distance 0) uniformly with how we already select test files during the traversal. However, there's a bit of a difference in terms of when we find out which file a symbol lives in for the roots vs. the symbols we find on the way. Because of that, the logic for distance 0 is implemented as a special case.

Reviewed By: mheiber

Differential Revision: D83848013

fbshipit-source-id: 241ca89160d2ace03710b6442c455f41106954a0