Add quickfixes for named parameter without `named`
Summary:
## What

Add quickfixes for errors introduced in {D83914745}. See snapshot tests

## Why

To provide a good user experience. Since we are introducing new syntax it may not otherwise be obvious to users how to fix the syntax error

## DevX Choice

The quickfix for *removing* the parameter name is good regardless of whether named parameters are enabled.

However, the quickfix for *adding* `named` isn't so useful until we expose named parameters.

I considered hiding the quickfix for adding `named` behind the unstable feature flag for named parameters, but:
- Seems like a fringe and temporary problem.
- I'm not so worried about the free advertising for the upcoming named parameters feature
- The plumbing is hard–right now it looks like we can only access unstable features from the parser errors environment

Reviewed By: andrewjkennedy

Differential Revision: D83914744

fbshipit-source-id: d0b3f9dd729f67275a732e894f92ee2730d34661