check parameter names in function types better
Summary:
Check parameter names in function types better: there shouldn't be the errors for "wrong arity" when it's the names that are wrong.
See snapshot tests for what I mean.

## Follow-on work

There's a bug where we miss type mismatches in a certain position, marked on a comment in this diff and fixed in this stack: {D83932882}

Reviewed By: andrewjkennedy

Differential Revision: D83914743

fbshipit-source-id: d62d81a72cfd18bc7b4590f54681fe78595e6d81