Improve error position for checking arguments against splat parameter
Summary: When checking arguments against a parameter type `...T` (a type splat), we generate a very poor position for the error in the arguments: the whole call-expression. Instead, generate a span of the remaining arguments that are being checked against the splat, if there are any, or the function expression, if there are no arguments left.

Reviewed By: vassilmladenov

Differential Revision: D83972017

fbshipit-source-id: 02f978a30d0869f03cbff1e068f0648ac5010c6c