extend Edenfs_watcher and hh_client telemetry
Summary:
This diff extends some telemetry collection related to file watching:

- In `ServerCheckStatus`, we now collect telemetry about how long `keep_trying_to_open` runs and how many iterations it takes.
- There's a new setting, `edenfs_file_watcher_report_telemetry`. If enabled, `Edenfs_watcher` will return additional telemetry. In `ServerNotifier` and `ClientCheckStatus`, we just attach this object to the ambient `Telemetry.t` object, if it exists.

Implementation note: Within `Edenfs_watcher`, I diverged from our usual "schema-less" approach to telemetry collection, instead defining records reflecting the exact shape of the data we collect.
This is helpful since we are exchanging some of this data between the worker thread and OCaml thread (see `AsyncTelemetry`), and I found it helpful to just use structs for everything else as well. This is not exposed to OCaml beyond the `Edenfs_watcher` API boundary: At that point we convert everything to the usual `Telemetry.t`.

Reviewed By: madgen

Differential Revision: D83995076

fbshipit-source-id: b11639b6977b44592e214ab1c194eed900cc8b48