Make IOBufShared->Bytes zero-copy if contiguous and unshared
Summary:
Previously, the conversion from `IOBufShared` to `bytes::Bytes` would always copy the IOBuf chain's contents into a `Vec<u8>` and then convert from `Vec` to `Bytes`.

This unnecessarily copied an arbitrarily large amount of bytes in the case that the IOBuf is contiguous and not shared, which is the common case. `Bytes` exposes a zero-copy constructor in the form of `Bytes::from_owner` for user-defined buffer types that are contiguous and immutable.

For the conversion from `IOBufShared` we are able to use this zero-copy constructor as long as the IOBuf is contiguous and unshared.

Reviewed By: Imxset21

Differential Revision: D83893898

fbshipit-source-id: 542ef3b0a1ae010b023759a2e5164c94fad2eccf