Fix OPAM build for Arm (#9644)
Summary:
Fetch the arm64 version of opam for installing deps, and disable the `ocaml-option-fp` flag which is not supported on aarch64.

The pinned versions of `dune` and `landmarks-ppx` still need updating to work on aarch64. This will be done as a followup since it seems to be causing issues in internal CI at the moment.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9644

Reviewed By: Wilfred

Differential Revision: D84053694

fbshipit-source-id: 7a2b9f746a613c55366992dbea0832b8bbf10523