Remove folly::to<> dependency from proxygen URLs
Summary:
Proxygen generally makes heavy use of Folly components.  However, some parts of Proxygenlib are also used by Tigon, which is trying to avoid Folly to reduce binary size.

The "fmt" library is generally recommended as a replacement for Folly formatting.  It has been incorporated into the C++ stdlib (although that is not available in our builds yet) and has a good reputation for producing small binaries.

Reviewed By: hanidamlaj, kvtsoy

Differential Revision: D81616974

fbshipit-source-id: 57e4c0a4b140fac9cfceba32e7e9b8f71118d7a3