Enable heterogeneous access through F14HashedKey
Summary:
Prior to this change, the following code would not compile:
```
void foo() {
  F14FastMap<std::string, std::string> myMap{};
  F14HashedKey<std::string_view> preHashedStringView{"Hello, world!"};
  myMap.find(preHashedStringView);
}
```

Therefore, we are forced to either change the type `preHashedStringView` to `F14HashedKey<std::string>`, making an unnecessary string copy, or to use the raw string view in the lookup, unnecessarily recalculating the key's hash.

This restriction is due to the `static_assert` statements in `computeHash` being overly restrictive, requiring the hashed key and the map to use identical 'hasher' types. In reality, we only need 'compatible' hasher types.

To resolve this, we add a new metaprogramming utility that treats subclasses of transparent range hashes as compatible. Probably there's a better way to check compatibility, but this seems to work well enough and can be expanded with additional cases if required. A new test case verifies that this case now compiles.

Reviewed By: skrueger

Differential Revision: D83560734

fbshipit-source-id: e04506b71a3daec7db5d2e1c79b6edf2ac323d7c