Fix Fake OOMs - Give request thread credit after stream payload is freed
Summary:
Problem -

In HHVM extensions, memory management is tightly coupled to the request thread's memory manager. If a background thread (not the request thread) frees memory that was originally allocated by the request thread, that memory is not credited back to the request's memory usage. This can cause memory accounting issues, such as the request appearing to use more memory than it actually does, and can make it difficult to enforce per-request memory limits or to debug memory leaks.

Solution -

1. Keep track of memory that was used by IOBuf constructed in request thread and later freed in IO thread.
2. Attach a callback to IOBuf free, so that when it is freed, it increments the above counter.
3. Whenever control is back in request thread (i.e. when next payload is ready to be sent), check if any memory was freed and give request thread credit by invoking `tl_heap->takeCreditForFreeOnOtherThread()`

This is similar to the fix we did for SR Client request-response few years back. (D27494143)

HHVM eng discussion post - https://fb.workplace.com/groups/165748143439038/permalink/32577107808542985/

There is still a delay between memory being freed and the call to `takeCreditForFreeOnOtherThread`. Also, it's possible that the previous payload hasn't been sent to the client yet. So, the payload size should be small enough that 2-3 of them can fit in memory. OOM can still occur in extreme cases.

Reviewed By: islamismailov

Differential Revision: D83105536

fbshipit-source-id: 9baa312da07a5dbbc6495daa71cce7e0210d941e