Remove hazptr dependency on Executor
Summary:
Hazptr is a low-level primitive, we should keep it as low-dependency as possible. In particular, all executor implementations depend on `RequestContext`, which in turn depends on hazptr. So if we wanted to make `RequestContext` support first class in `Executor`, we would create a cyclic dependency.

It's better to have `Hazptr` not depend on `Executor` at all, since it doesn't really need it, it just needs something that looks like `Executor::add()`, and we can make it succeed conditionally on whether we actually have an executor available (one was registered and we're not at shutdown), which simplifies the handling of the inline fallback.

Reviewed By: yfeldblum

Differential Revision: D82303243

fbshipit-source-id: 720ef233d2078004b82b8de12788ab1594d98e9e