Use sealed list for classes when computing disjointness
Summary:
Currently we consider the following case type to be invalid because the variants are overlapping:

```
interface I {}

case type MyCaseType = I | Awaitable<I>;
```

This feels incorrect because Awaitable is sealed with all leaf classes being final.

The issue is we only leverage the sealed list for interfaces. This looks like an oversight and don't see any reason not to apply it to classes as well.

Reviewed By: viratyosin

Differential Revision: D84212595

fbshipit-source-id: 4668aa818714c73ad2032e6ec090a387d5bcec4a