Parser: pass in string mode as ctor arg
Summary:
I want to specify which parser strategy to use per-connection of defaulting to one for an entire binary. Start by adding the parser strategy as an arg to Parser ctor.

No functional change in this diff.

Parser ctor specifically takes an rvalue ref since in 99.9% of cases the arg is passed in as an rvalue from THRIFT_FLAG() in the caller.

Reviewed By: robertroeser

Differential Revision: D84029705

fbshipit-source-id: 6b7dae577203976e6eee8e3b23d6526278d32c62