Parser: add a weak func that returns strategy based on socket
Summary:
I want to specify which parser strategy to use per-connection instead of defaulting to one for an entire binary. Add an overridable function `getSocketParser` that takes an AsyncTransport and returns a parser strategy as a string.

The default impl of this overridable function returns the binary wide strategy specified in `rocket_frame_parser`. There is no functional change in this diff.

Added a unit test that demonstrates how a service binary might override this behaviour.

Changes in V4:

Added another ctor for Parser that takes a `const std::string&` that eliminates any unnecessary copying. `THRIFT_FLAG()` returns a `std::string` by value and as a prvalue will bind to `std::string&&`. The new `cfg.parserStrategy` will bind to `const std::string&`. It's worth noting that the previous implementation `detail::stringToMode(THRIFT_FLAG(rocket_frame_parser))` always resulted in a copy due to stringToMode taking a `const std::string&`.

Reviewed By: robertroeser

Differential Revision: D84032488

fbshipit-source-id: 65014c9eed3a3dccdebde4d27a5f973326e51355