folly/python/request_context - perf boost
Summary:
Reading from the PyContext on every setContext taking place in a python thread is slow.  So what if we use a thread_local to store the last_set_context.  It several fewer operations and does not involve python c-api

This will avoid the setContextWatcher setting the already set conext after we call setContext from the PyContext Watcher

Reviewed By: yfeldblum

Differential Revision: D84313806

fbshipit-source-id: 8c70ddf19dda9d646ce38ed277a1275b5e526c2f