Setup an experiment to avoid doing unnecessary work relevant only to save state creation during regular type checks
Summary:
When hh falls back on hh_distc it unconditionally sets following options:
https://www.internalfb.com/code/fbsource/[af3574829eefaa1e94883e7996304d54f69d3465]/fbcode/hphp/hack/src/facebook/hh_distc/check_ffi/hh_distc_rust_ffi.rs?lines=121-122
While if called directly these options require `--save-state` to be passed as well:
https://www.internalfb.com/code/fbsource/[af3574829eefaa1e94883e7996304d54f69d3465]/fbcode/hphp/hack/src/facebook/hh_distc/check/typecheck.rs?lines=156-157%2C166-167
Without providing `--save-state` calling hh_distc with `--save-64bit` or `--write-compressed-depgraph` would result in an error.
This diff sets up an experiment to see what impact avoiding this unnecessary work would make. After the experiment conclusion we'll make this behaviour default to make sure that it is the same for both FFI and CLI invocations of hh_distc.

Reviewed By: madgen

Differential Revision: D84265139

fbshipit-source-id: 6b8ad12c321ed63adeae05e059f9a6b84be03adc