benchmark for hazptr_domain::load_hazptr_vals
Summary:
A new benchmark binary showing some common scenarios. By no means complete.

In particular, includes measurement of the cost of walking the hprec's during reclamation for various lengths of the hprec-list.

Note that in the short linked-list cases, the asymmetric-thread-fence-heavy dominates (from `perf`). We start to see the effect of the scattered linked-list walk when the length gets into the thousands.

Reviewed By: skrueger

Differential Revision: D83933375

fbshipit-source-id: 0c03be62f6049c98fe9528a23bf0a763d654ca8c