Flag function pointers in destructor table
Summary:
When a constructor takes the address of some function and
manually attributes it to some memory location, no data relocations
are emitted, and the references are hardcoded in the constructor code
itself. ROAR will miss those and only the optimized version of this
constructor will have references that bypasses ROAR trampolines, but
since this is a constructor, it will only run once and never have an
optimized version. In this case, we manually flag to ROAR that we have
functions pointers in the application memory (in this case, global
space), so ROAR can patch those locations with updated versions of
these destructors.

When profiling HHVM on ROAR with LBRs, we can see that these are the
top trampoline usages.

Reviewed By: ottoni

Differential Revision: D84309068

fbshipit-source-id: b7493bf784f35fcfd6ceed1fa3e57c1462f2499b