Support virtual inheritance paths to std::exception in exception tracer library
Summary:
This diff implements stack trace collection for derived exceptions in virtual inheritance hierarchies.

## Description of current behavior

Stack traces for exceptions `A` at the time they are thrown are stored by `SmartExceptionTracer` in a global `map<void*, ExceptionInfo>`. Previously, when performing lookups to obtain stack traces, the address of the object `B` to obtain a stack trace for needed to match the pointer previously stored. This works for simple cases where `A` and `B` are [pointer interconvertible](https://en.cppreference.com/w/cpp/language/static_cast.html#pointer-interconvertible), but failed lookups for cases where there is a non-zero offset between addresses of B and the thrown object A due to their differing layout.

## Fix

The fix uses `dynamic_cast<void*>` to obtain the address of the [most derived object](https://eel.is/c++draft/intro.object#6) associated with the object passed into `getTrace` for the case where a reference to a derived type (like `std::exception`) is passed in. This ensures that the address used in the lookup in the global map of thrown exceptions corresponds to the address placed there by the hook (ie the address passed by the runtime to `__cxa_throw()`.

## Side effects of this change

One side effect of how the fix is implemented is that we avoid a `dynamic_cast` to `std::exception` in the lookup paths where `std::exception_ptr` or `folly::ExceptionWrapper` are passed. This enables support for collecting stack traces for thrown objects that are not derived from `std::exception` (see the changed `NonStdException` test case).

In addition to that, this diffs skips unnecessary RTTI inspection of thrown objects to find out whether they are compatible with `std::exception` as the pointer to the originally thrown exception is already available when `std::exception_ptr` or `folly::exception_wrapper` is available, which increases performance slightly (see benchmarks below).

Reviewed By: shr, yfeldblum

Differential Revision: D84131474

fbshipit-source-id: 4e5434ffa01ad5739fed4226ec5b8891ee3c2daa