fix absolute vs relative path bug at distance 0
Summary:
This diff fixes a bug in the logic for checking whether any of the root symbols given to `hh --find-my-tests` is inside a test file. It should return an absolute path, but it was returning a relative one.

My test suite didn't catch this because it was only checking the file names, not paths, in the output of `hh --find-my-tests`. I've updated the test suite to compare full paths now.

Reviewed By: madgen

Differential Revision: D84529707

fbshipit-source-id: 4d050196649ba17276e21daa38e382ab73e3fb20