Simplify thrift scheduler arguments
Summary:
This diff aims to simplify how we configure the thrift scheuduler. We now have 3 scheduler implementations, thread-pool, fork-join-pool, and virtual thread scheduler. It is also possible to have a client scheduler. To confiugre these in the past it was quite confusing, today we use 3 flags to confgure the three options.

(1) Configure which scheduler type you want
```
-Dthrift.off-loop-scheduler=[ thread-pool | fork-join-pool | virtual-thread ]
```

(2) The target concurrency level for the off loop schduler, this will corespond to the max number of threads for the fork-join-pool and Thread Pool Executor, but will only limit concurrency for virtual threads if the `thrift.virtual-thread-limit-concurrency` flag is specified.
```
-Dthrift.scheduler-concurrency.count=[ int ]
```

(3) The target concurrency level for the client off loop scheduler. Similar to (2) but for the client scheduler if `thrift.separate-offloop-scheduler=true`.

```
-Dthrift.client-scheduler-concurrency.count=[ int ]
```

Reviewed By: adolfojunior, RayanRal

Differential Revision: D84387127

fbshipit-source-id: e3c49c38a8f6a4b518dfc14025d23eb4a1f1e878