Change error message strategy
Summary:
Previously we were trying to process the error callbacks and reasons after sub-typing. However, we didn't have enough information to reliably determine when a reason projection referred to the method call argument, versus some other call that is part of the provenance.

Now we will update the reason callback inside of subtyping each time we descend into an argument ahead of time

Reviewed By: andrewjkennedy

Differential Revision: D84158655

fbshipit-source-id: 6b474a9bb94f2493ad44275a8e31cbfdf9d7cbbd