Packages in Facts
Summary: We want to have Facts be able to provide the package for any given path.  To do this, we need to know about any package overrides which might be in the file, so we need to pipe that from the decl parser through to the FFI so Facts can grab it.

Reviewed By: nt591

Differential Revision: D82849959

fbshipit-source-id: 2c06400f2dfc18300b7d1721983f66ade06a24fc