move writing batching per eventbase
Summary:
In RocketServerConnection batching occurred per connection, but this moves the batching to occur the per eventbase instances.

The previous version of rocketserverconnection  batched per connection. The problem with this is if the frequency of requests was low enough per connection it would cause batching to not be effective, it won't batch and it will send one item at time. Moving it per eventbase tick means connections will now batch there requests together encouraging more batching to occur. For example if in the same event loop tick there were 4 requests from separate connections - it would call write to the async socket 4 times. Now, instead of call 4 times it will calling it one time on the tick unless a threshold has been met.

Reviewed By: cevans87

Differential Revision: D83917443

fbshipit-source-id: edbba9e935bee999b949130745903657f98a4d03